Perform expansions in separator
so later prompt expansions can be handled in the prompt itself.

E.g. a separator format like `%F{%(!.${COLOR_ROOT}.${COLOR_USER})/%f`
would be expanded to `%F{${COLOR_USER}}/%f` by the prompt, but not
further than that. So the separator needs to be expanded to
`%F{%(!.red.cyan)/%f` first.